prefCode から prefName に変換するロジックをprefecture.ts に移動した
- これまでの修正でグラフデータの生成ロジックが prefCode から prefName に変換するロジックへの興味を失ったため
