+++ src/domain/prefecture.ts
@@ -6,3 +6,8 @@ export type Prefecture = {
   prefCode: number
   prefName: string
 }
+
+export function convertPrefCodeIntoPrefName(prefCode: number, prefecturesData: PrefecturesData): string {
+  const targetData = prefecturesData.data.find((d) => d.prefCode === prefCode)
+  return targetData ? targetData.prefName : ''
+}
