+++ src/domain/prefecturesPopulationGraphByYear.spec.ts
@@ -1,8 +1,4 @@
-import {
-  convertFromPrefecturesPopulationData,
-  genYearsSortedInAsc,
-  convertPrefCodeIntoPrefName,
-} from './prefecturesPopulationGraphByYear'
+import { convertFromPrefecturesPopulationData, genYearsSortedInAsc } from './prefecturesPopulationGraphByYear'
 
 const samplePrefecturesPopulationData = {
   data: [
@@ -41,21 +37,6 @@ const samplePrefecturesPopulationData = {
           value: 3,
         },
       ],
-    },
-  ],
-  isLoading: false,
-  isError: false,
-}
-
-const samplePrefecturesData = {
-  data: [
-    {
-      prefCode: 0,
-      prefName: '新潟県',
-    },
-    {
-      prefCode: 1,
-      prefName: '山形県',
     },
   ],
   isLoading: false,
@@ -159,19 +140,6 @@ describe('genSortedYears', () => {
         ],
       }
       expect(genYearsSortedInAsc(prefecturesPopulationData)).toEqual([1992, 2020, 2025, 2030])
-    })
-  })
-})
-
-describe('convertPrefCodeIntoPrefName', () => {
-  describe('prevCode が PrefecturesData に入っていること', () => {
-    it('都道府県名が返ってくること', () => {
-      expect(convertPrefCodeIntoPrefName(1, samplePrefecturesData)).toEqual('山形県')
-    })
-  })
-  describe('prevCode が PrefecturesData に入っていない', () => {
-    it('空文字列が返ってこないこと', () => {
-      expect(convertPrefCodeIntoPrefName(3, samplePrefecturesData)).toEqual('')
     })
   })
 })
